fix(useCopy): refactor, move to shared composable directory
